libten_tusscher_2004_epi_S1_tr6_pop69.so have the same code now..

+++ src/models_library/ten_tusscher_2004_RS_CPU_epi_S1_tr6_pop69.c
@@ -347,7 +347,7 @@ real GpK=0.0199551557341385;
 A=0.00934661872479498*CaSRsquare/(0.0625f+CaSRsquare)+0.00479306991518623;
     Irel=A*sd*sg;
    // Ileak=0.00008f*(CaSR-Cai);
-Ileak=5.60468833850934e-05*(CaSR-Cai);
+    Ileak=5.60468833850934e-05*(CaSR-Cai);
     SERCA=Vmaxup/(1.f+(Kupsquare/Caisquare));
     CaSRCurrent=SERCA-Irel-Ileak;
     CaCSQN=Bufsr*CaSR/(CaSR+Kbufsr);
@@ -359,7 +359,7 @@ Ileak=5.60468833850934e-05*(CaSR-Cai);
     dCai=dt*(CaCurrent-CaSRCurrent);
     bc=Bufc-CaBuf-dCai-Cai+Kbufc;
     cc=Kbufc*(CaBuf+dCai+Cai);
-    Cai=(sqrt(bc*bc+4*cc)-bc)/2;
+    Cai=(sqrtf(bc*bc+4*cc)-bc)/2;
